Some more work with bug fixing.
Assumption that chiled bounds are always smoller that the parents bounds is false.
Have to think about it...

+++ quick_benchmark.h
@@ -98,14 +98,14 @@ unsigned long long __rdtsc(void);
 #define QB_BENCH_WITH_NAME(file, n_runs, x, display, ...) \
 	do { \
 		uint64_t _tot = 0; \
-		for (int _j = 0; _j < n_runs; _j++) { \
+		for (uint64_t _j = 0; _j < n_runs; _j++) { \
 			uint64_t _start, _end;	\
 			_start = (uint64_t)QB_SETTINGS_BENCH_FUNC(); \
 			x(__VA_ARGS__); \
 			_end = (uint64_t)QB_SETTINGS_BENCH_FUNC(); \
 			_tot += _end - _start; \
 		} \
-		fprintf(file, "%-50s calls: %.4d %10s %llu.%llu ms\n", \
+		fprintf(file, "%-50s calls: %.4d %10s %lu.%lu ms\n", \
 			display, (int)n_runs, "time:", _tot / n_runs / 1000000, _tot / n_runs % 1000000 / 1000); \
 	} while(0)
